Add Session

+++ resources/views/nilai/index.blade.php
@@ -3,9 +3,9 @@
     <center>
         <b>
             <h2>LIST DATA NILAI</h2>
-            {{-- @if (session('user')->role == 'guru') --}}
+            @if (session('user')->role == 'guru')
                 <a href="/nilai/create" class="button-primary">Tambah Data Nilai</a>
-            {{-- @endif --}}
+            @endif
             @if (session('success'))
                 <p class="text-success">{{ session('success') }}</p>
             @endif
@@ -22,9 +22,9 @@
                     <th>UTS</th>
                     <th>UAS</th>
                     <th>NA</th>
-                    {{-- @if (session('user')->role == 'guru') --}}
+                    @if (session('user')->role == 'guru')
                         <th>ACTION</th>
-                    {{-- @endif --}}
+                    @endif
                 </tr>
                 @foreach ($nilai as $each)
                     <tr>
@@ -35,12 +35,12 @@
                         <td>{{ $each->uts }}</td>
                         <td>{{ $each->uas }}</td>
                         <td>{{ $each->na }}</td>
-                        {{-- @if (session('user')->role == 'guru') --}}
+                        @if (session('user')->role == 'guru')
                             <td>
                                 <a href="/nilai/edit/{{ $each->id }}" class="button-warning">EDIT</a>
                                 <a href="/nilai/destroy/{{ $each->id }}" class="button-danger" onclick="return alert('Yakin Hapus?')">HAPUS</a>
                             </td>
-                        {{-- @endif --}}
+                        @endif
                     </tr>
                 @endforeach
             </table>
